format and restructuure
