+++ packages/vector_graphics_compiler/lib/src/svg/resolver.dart
@@ -1,0 +1,283 @@
+import '../geometry/basic_types.dart';
+import '../geometry/matrix.dart';
+import '../geometry/path.dart';
+import '../paint.dart';
+import 'node.dart';
+import 'parser.dart';
+import 'visitor.dart';
+
+/// A visitor class that processes relative coordinates in the tree into a
+/// single coordinate space, removing extra attributes, empty nodes, resolving
+/// references/masks/clips.
+class ResolvingVisitor extends Visitor<Node> {
+  late Rect _bounds;
+  final List<AffineMatrix> _transformStack = <AffineMatrix>[
+    AffineMatrix.identity,
+  ];
+
+  @override
+  Node visitClipNode(ClipNode clipNode) {
+    final List<Path> transformedClips = <Path>[
+      for (Path clip in clipNode.resolver(clipNode.clipId))
+        clip.transformed(_transformStack.last)
+    ];
+    if (transformedClips.isEmpty) {
+      return clipNode.child.accept(this);
+    }
+    return ResolvedClipNode(
+      clips: transformedClips,
+      child: clipNode.child.accept(this),
+    );
+  }
+
+  @override
+  Node visitMaskNode(MaskNode maskNode) {
+    final AttributedNode? resolvedMask = maskNode.resolver(maskNode.maskId);
+    if (resolvedMask == null) {
+      return maskNode.child.accept(this);
+    }
+    final Node child = maskNode.child.accept(this);
+    final AffineMatrix childTransform =
+        maskNode.concatTransform(_transformStack.last);
+
+    _transformStack.add(childTransform);
+    final Node mask = resolvedMask.accept(this);
+    _transformStack.removeLast();
+
+    return ResolvedMaskNode(
+      child: child,
+      mask: mask,
+      blendMode: maskNode.blendMode,
+    );
+  }
+
+  @override
+  Node visitParentNode(ParentNode parentNode) {
+    final AffineMatrix current = _transformStack.last;
+    final AffineMatrix nextTransform = parentNode.concatTransform(current);
+    _transformStack.add(nextTransform);
+
+    final Paint? saveLayerPaint = parentNode.createLayerPaint();
+    final Node result;
+    if (saveLayerPaint == null) {
+      result = ParentNode(
+        SvgAttributes.empty,
+        precalculatedTransform: AffineMatrix.identity,
+        children: <Node>[
+          for (Node child in parentNode.children) child.accept(this),
+        ],
+      );
+    } else {
+      result = SaveLayerNode(
+        SvgAttributes.empty,
+        paint: saveLayerPaint,
+        children: <Node>[
+          for (Node child in parentNode.children) child.accept(this),
+        ],
+      );
+    }
+    _transformStack.removeLast();
+    return result;
+  }
+
+  @override
+  Node visitPathNode(PathNode pathNode) {
+    final AffineMatrix transform =
+        _transformStack.last.multiplied(pathNode.attributes.transform);
+    final Path transformedPath = pathNode.path.transformed(transform);
+    final Rect originalBounds = pathNode.path.bounds();
+    final Rect newBounds = transformedPath.bounds();
+    final Paint? paint = pathNode.computePaint(originalBounds, transform);
+    if (paint != null) {
+      return ResolvedPathNode(
+        paint: paint,
+        bounds: newBounds,
+        path: transformedPath,
+      );
+    }
+    return Node.empty;
+  }
+
+  @override
+  Node visitTextNode(TextNode textNode) {
+    final Paint? paint = textNode.computePaint(_bounds, _transformStack.last);
+    final TextConfig textConfig =
+        textNode.computeTextConfig(_bounds, _transformStack.last);
+    if (paint != null && textConfig.text.trim().isNotEmpty) {
+      return ResolvedTextNode(
+        textConfig: textConfig,
+        paint: paint,
+      );
+    }
+    return Node.empty;
+  }
+
+  @override
+  Node visitViewportNode(ViewportNode viewportNode) {
+    _bounds = Rect.fromLTWH(0, 0, viewportNode.width, viewportNode.height);
+    _transformStack.add(viewportNode.transform);
+    return ViewportNode(
+      SvgAttributes.empty,
+      width: viewportNode.width,
+      height: viewportNode.height,
+      transform: AffineMatrix.identity,
+      children: <Node>[
+        for (Node child in viewportNode.children) child.accept(this),
+      ],
+    );
+  }
+
+  @override
+  Node visitDeferredNode(DeferredNode deferredNode) {
+    final AttributedNode? resolvedNode =
+        deferredNode.resolver(deferredNode.refId);
+    if (resolvedNode == null) {
+      return Node.empty;
+    }
+    final AttributedNode concreteRef =
+        resolvedNode.applyAttributes(deferredNode.attributes);
+    return concreteRef.accept(this);
+  }
+
+  @override
+  Node visitEmptyNode(Node node) => node;
+
+  @override
+  Node visitResolvedText(ResolvedTextNode textNode) {
+    assert(false);
+    return textNode;
+  }
+
+  @override
+  Node visitResolvedPath(ResolvedPathNode pathNode) {
+    assert(false);
+    return pathNode;
+  }
+
+  @override
+  Node visitResolvedClipNode(ResolvedClipNode clipNode) {
+    assert(false);
+    return clipNode;
+  }
+
+  @override
+  Node visitResolvedMaskNode(ResolvedMaskNode maskNode) {
+    assert(false);
+    return maskNode;
+  }
+
+  @override
+  Node visitSaveLayerNode(SaveLayerNode layerNode) {
+    assert(false);
+    return layerNode;
+  }
+}
+
+/// A block of text that has its position and final transfrom fully known.
+///
+/// This should only be constructed from a [TextNode] in a [ResolvingVisitor].
+class ResolvedTextNode extends Node {
+  /// Create a new [ResolvedTextNode].
+  ResolvedTextNode({
+    required this.textConfig,
+    required this.paint,
+  });
+
+  /// The text configuration to draw this piece of text.
+  final TextConfig textConfig;
+
+  /// The paint used to draw this piece of text.
+  final Paint paint;
+
+  @override
+  S accept<S>(Visitor<S> visitor) {
+    return visitor.visitResolvedText(this);
+  }
+
+  @override
+  void visitChildren(NodeCallback visitor) {}
+}
+
+/// A path node that has its bounds fully computed.
+///
+/// This should only be constructed from a [PathNode] in a [ResolvingVisitor].
+class ResolvedPathNode extends Node {
+  /// Create a new [ResolvedPathNode].
+  ResolvedPathNode({
+    required this.paint,
+    required this.bounds,
+    required this.path,
+  });
+
+  /// The paint for the current path node.
+  final Paint paint;
+
+  /// The bounds estimate for the current path.
+  final Rect bounds;
+
+  /// The path to be drawn.
+  final Path path;
+
+  @override
+  S accept<S>(Visitor<S> visitor) {
+    return visitor.visitResolvedPath(this);
+  }
+
+  @override
+  void visitChildren(NodeCallback visitor) {}
+}
+
+/// A clip node where all paths are known and transformed in a single
+/// coordinate space.
+///
+/// This should only be constructed from a [ClipNode] in a [ResolvingVisitor].
+class ResolvedClipNode extends Node {
+  /// Create a new [ResolvedClipNode].
+  ResolvedClipNode({
+    required this.clips,
+    required this.child,
+  });
+
+  /// One or more clips to apply to rendered children.
+  final List<Path> clips;
+
+  /// The child node.
+  final Node child;
+
+  @override
+  S accept<S>(Visitor<S> visitor) {
+    return visitor.visitResolvedClipNode(this);
+  }
+
+  @override
+  void visitChildren(NodeCallback visitor) {}
+}
+
+/// A mask node with child and mask fully resolved.
+///
+/// This should only be constructed from a [MaskNode] in a [ResolvingVisitor].
+class ResolvedMaskNode extends Node {
+  /// Create a new [ResolvedMaskNode].
+  ResolvedMaskNode({
+    required this.child,
+    required this.mask,
+    required this.blendMode,
+  });
+
+  /// The child to apply as a mask.
+  final Node mask;
+
+  /// The child of this mask layer.
+  final Node child;
+
+  /// The blend mode to apply when saving a layer for the mask, if any.
+  final BlendMode? blendMode;
+
+  @override
+  S accept<S>(Visitor<S> visitor) {
+    return visitor.visitResolvedMaskNode(this);
+  }
+
+  @override
+  void visitChildren(NodeCallback visitor) {}
+}
